+++ common/src/main/java/club/iananderson/seasonhud/client/gui/components/sliders/rgb/RgbSlider.java
@@ -5,8 +5,9 @@ import club.iananderson.seasonhud.client.gui.components.sliders.BasicSlider;
 import club.iananderson.seasonhud.config.Config;
 import club.iananderson.seasonhud.impl.seasons.Seasons;
 import club.iananderson.seasonhud.util.Rgb;
+import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.ChatFormatting;
-import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.client.Minecraft;
 import net.minecraft.network.chat.Component;
 import org.jetbrains.annotations.NotNull;
 
@@ -51,13 +52,13 @@ public class RgbSlider extends BasicSlider {
   }
 
   @Override
-  public void renderWidget(@NotNull GuiGraphics graphics, int mouseX, int mouseY, float partialTick) {
+  public void renderBg(@NotNull PoseStack graphics, Minecraft mc, int mouseX, int mouseY) {
     if (!enableColor) {
       this.active = false;
       this.isHovered = false;
     }
 
-    super.renderWidget(graphics, mouseX, mouseY, partialTick);
+    super.renderBg(graphics, mc, mouseX, mouseY);
   }
 
   @Override