+++ common/src/main/java/club/iananderson/seasonhud/client/gui/screens/ColorScreen.java
@@ -18,7 +18,6 @@ import java.util.Arrays;
 import java.util.EnumSet;
 import java.util.List;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.AbstractWidget;
 import net.minecraft.client.gui.components.CycleButton;
 import net.minecraft.client.gui.components.EditBox;