+++ common/src/main/java/club/iananderson/seasonhud/util/DrawUtil.java
@@ -28,32 +28,32 @@ public class DrawUtil {
     RenderSystem.setShaderTexture(0, texture);
     guiComponent.blit(poseStack, x, y, u, v, leftBorder, topBorder);
     guiComponent.blit(poseStack, x + leftBorder + canvasWidth, y, u + leftBorder + fillerWidth, v, rightBorder,
-        topBorder);
+                      topBorder);
     guiComponent.blit(poseStack, x, y + topBorder + canvasHeight, u, v + topBorder + fillerHeight, leftBorder,
-                     bottomBorder);
+                      bottomBorder);
     guiComponent.blit(poseStack, x + leftBorder + canvasWidth, y + topBorder + canvasHeight,
-        u + leftBorder + fillerWidth, v + topBorder + fillerHeight, rightBorder, bottomBorder);
+                      u + leftBorder + fillerWidth, v + topBorder + fillerHeight, rightBorder, bottomBorder);
 
     int i;
     for (i = 0; i < xPasses + (remainderWidth > 0 ? 1 : 0); ++i) {
       guiComponent.blit(poseStack, x + leftBorder + i * fillerWidth, y, u + leftBorder, v,
-                       i == xPasses ? remainderWidth : fillerWidth, topBorder);
+                        i == xPasses ? remainderWidth : fillerWidth, topBorder);
       guiComponent.blit(poseStack, x + leftBorder + i * fillerWidth, y + topBorder + canvasHeight, u + leftBorder,
-                       v + topBorder + fillerHeight, i == xPasses ? remainderWidth : fillerWidth, bottomBorder);
+                        v + topBorder + fillerHeight, i == xPasses ? remainderWidth : fillerWidth, bottomBorder);
 
       for (int j = 0; j < yPasses + (remainderHeight > 0 ? 1 : 0); ++j) {
         guiComponent.blit(poseStack, x + leftBorder + i * fillerWidth, y + topBorder + j * fillerHeight, u + leftBorder,
-                         v + topBorder, i == xPasses ? remainderWidth : fillerWidth,
-                         j == yPasses ? remainderHeight : fillerHeight);
+                          v + topBorder, i == xPasses ? remainderWidth : fillerWidth,
+                          j == yPasses ? remainderHeight : fillerHeight);
       }
     }
 
     for (i = 0; i < yPasses + (remainderHeight > 0 ? 1 : 0); ++i) {
       guiComponent.blit(poseStack, x, y + topBorder + i * fillerHeight, u, v + topBorder, leftBorder,
-                       i == yPasses ? remainderHeight : fillerHeight);
+                        i == yPasses ? remainderHeight : fillerHeight);
       guiComponent.blit(poseStack, x + leftBorder + canvasWidth, y + topBorder + i * fillerHeight,
-                       u + leftBorder + fillerWidth, v + topBorder, rightBorder,
-                       i == yPasses ? remainderHeight : fillerHeight);
+                        u + leftBorder + fillerWidth, v + topBorder, rightBorder,
+                        i == yPasses ? remainderHeight : fillerHeight);
     }
 
   }