+++ forge/src/main/java/club/iananderson/seasonhud/impl/curios/CuriosCompat.java
@@ -1,11 +1,12 @@
 package club.iananderson.seasonhud.impl.curios;
 
 import club.iananderson.seasonhud.impl.curios.item.CuriosCalendar;
-import club.iananderson.seasonhud.platform.Services;
 import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.fml.InterModComms;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-import top.theillusivec4.curios.api.CuriosApi;
+import top.theillusivec4.curios.api.SlotTypeMessage;
+import top.theillusivec4.curios.api.SlotTypePreset;
 
 public class CuriosCompat {
   public CuriosCompat() {
@@ -13,8 +14,14 @@ public class CuriosCompat {
     eventBus.addListener(this::setup);
   }
 
+  public static void registerSlots() {
+    InterModComms.sendTo("curios", SlotTypeMessage.REGISTER_TYPE, () -> SlotTypePreset.CHARM.getMessageBuilder()
+        .build());
+  }
+
   public void setup(final FMLCommonSetupEvent evt) {
-    CuriosApi.registerCurio(Services.SEASON.calendar(), new CuriosCalendar());
+    CuriosCalendar.initCapabilities();
+    CuriosCompat.registerSlots();
   }
 
 }