+++ forge/src/main/java/club/iananderson/seasonhud/impl/curios/item/CuriosCalendar.java
@@ -1,15 +1,40 @@
 package club.iananderson.seasonhud.impl.curios.item;
 
+import club.iananderson.seasonhud.platform.Services;
+import javax.annotation.Nonnull;
+import net.minecraft.core.Direction;
 import net.minecraft.world.item.ItemStack;
-import top.theillusivec4.curios.api.SlotContext;
+import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.common.capabilities.ICapabilityProvider;
+import net.minecraftforge.common.util.LazyOptional;
+import org.jetbrains.annotations.Nullable;
+import top.theillusivec4.curios.api.CuriosCapability;
+import top.theillusivec4.curios.api.type.capability.ICurio;
 import top.theillusivec4.curios.api.type.capability.ICurioItem;
 
 public class CuriosCalendar implements ICurioItem {
   public CuriosCalendar() {
   }
 
-  @Override
-  public void curioTick(SlotContext slotContext, ItemStack stack) {
-    // ticking logic here
+  public static ICapabilityProvider initCapabilities() {
+    ICurio curio = new ICurio() {
+      final ItemStack stack = new ItemStack(Services.SEASON.calendar().asItem());
+
+      @Override
+      public ItemStack getStack() {
+        return stack;
+      }
+
+    };
+    return new ICapabilityProvider() {
+      private final LazyOptional<ICurio> curioOpt = LazyOptional.of(() -> curio);
+
+      @Nonnull
+      @Override
+      public <T> LazyOptional<T> getCapability(@Nonnull Capability<T> cap, @Nullable Direction side) {
+
+        return CuriosCapability.ITEM.orEmpty(cap, curioOpt);
+      }
+    };
   }
 }